[Ionut] - More Linux port work
git-svn-id: https://xp-dev.com/svn/Divide-Framework/trunk@458 e89c6373-eeba-4702-abef-d26284fe765c

Former-commit-id: 9546de99906babf3d299f18b0b2beb9faafc0a46 [formerly 5d7297a29bbb7135e613e8802376ec12156149af] [formerly 1b9c49fd811ef4e3f0c3696590b38563df52de56 [formerly da447ad66816dcfa218c9f3a14bea8fd3dc0976b]]
Former-commit-id: ff56c6c05e32dab914ef0193b52df47fc66bbae7 [formerly b327c7dfe0f33fbebc5632cbc34992550ec22d67]
Former-commit-id: dfbeb829f012ba8fdf40c674bafcc3a3d74ff861

+++ Source Code/AI/Sensors/Headers/VisualSensor.h
@@ -67,7 +67,7 @@ class VisualSensor : public Sensor {
     inline F32 getDistanceToNode(U32 containerID, U64 nodeGUID) {
         F32 distanceSq = getDistanceToNodeSq(containerID, nodeGUID);
         if (distanceSq < std::numeric_limits<F32>::max() - 1.0f) {
-            return std::sqrtf(distanceSq);
+            return std::sqrt(distanceSq);
         }
         return distanceSq;
     }
