+++ Source Code/Physics/PhysX/Actors/pxShapeScaling.cpp
@@ -120,7 +120,7 @@ void Scale(physx::PxCapsuleGeometry &geometry, physx::PxTransform &pose,
     geometry.halfHeight *= scaleMat.column0.magnitude();
     // Bi-directional shape-space approximation
     geometry.radius *=
-        std::sqrtf(scaleMat.column1.magnitude() * scaleMat.column2.magnitude());
+        std::sqrt(scaleMat.column1.magnitude() * scaleMat.column2.magnitude());
 
     pose.p = pose.p.multiply(scaling);
 }
